MarketStreaming/code/containers/retrievemarketdata/requirements.txt
pandas
requests
azure-eventhub
datetime